DriverOI and OperatorOI changes
Eliminate getLeftForward and getRightForward in both classes, this is old code we developed (I think in 2023 based on the scaling I see) and they do not seem to serve any current purpose.

All controls can probably use a look through. Do another pass through with Philip of the 2 driver controls spreadsheet and get things to match.  Make sure we haven't forgotten things that we want on there.  There's some stuff in the code that seems out-of-date, both including code that's commented out, as well as code that is not commented out.   The only commented out code should be any sort of thing that we specifically do want disabled right now, such as a command that is unsafe to run, or a special binding for testing a command temporarily. Otherwise, DELETE/EDIT.