standardized elevator position (no more height))
getElevatorHeight - are we actually implementing elevator height, or can we just work with the native cancoder units and develop presets?  (This is similar philosophically to what I said in the Arm, but maybe even more relevant as I doubt we care about what these units mean - do we need actual height in meters/inches for any reason? Adds complexity).
instead of "isAtHeight", maybe refactor this to an appropriate "isAtPosition" method,  assuming we do not do a "height" as mentioned in (a)