Implemented sensortomechanism and rotortosensor ratios for elevator
Related to (e), unless it's intentional, we are missing code for the sensor for the Rotor to Sensor and Sensor to Mechanism ratios for our elevator, I haven't spent much time thinking through this or checking on the cancoder setup. We should discuss/check on this?  Actually I don't know what sensor to mechanism would exactly even mean for an elevator, compared to the output shaft on a rotating arm???  (Maybe one of you has thought about all of this already? I didn’t do the research on best practices for elevators).