Implemented Auto PID constants + renamed OperatorConstants
utoConstants: These do not appear actually implemented in the code, we should set these constants to the hard-coded values in Autonomous.java and then keep the TODO in case we need to tune further

OperatorConstants: should be renamed from OperatorConstants to OIConstants(?).  Missing ports for both driver and operator.  Currently all these are hard-coded into the DriverOI and OperatorOI classes. (Note: philosophically we also could keep the hard-coding in this case and omit these as they are not numbers we would ever change).