added java docs
Need some updating on javadoc comments, we've fallen behind on this / all programmers need to be aware of our coding standards for implementing a new method.
EXCEPTION/CLARIFICATION: if it is a completely self-explanatory method such as 'onlyLeftTriggerHeld' in DriverOI.java, which has only 1 line of code which is unambiguously clear, we do not need to do a javadoc comment. Don’t waste time on this. However, Javadoc comments ALWAYS need to exist for core subsystem methods and especially for any methods involving any type of units or serving specialized purposes. If in doubt, write a Javadoc comment.