Added charuco calibration folder
Charuco calibration folder: There is a "LL3G_ChArUco_Calibration.json" in a random directory level that is non-specific and also should be organized into a folder if it is something current we need.  We do need to make sure we have a directory (I suggest this is at the same directory level as "ElasticPresets", perhaps?) to store charuco calibrations.